updated tests - still in progress
