+++ test/test_phipdata.py
@@ -24,15 +24,55 @@ from sim_test_generator import generate_sim_ds
 
 
 # local dependency
-from phippery.phipdata import counts_metadata_to_dataset
-from phippery.phipdata import collect_sample_metadata
-from phippery.phipdata import collect_peptide_metadata
+from phippery.phipdata import convert_peptide_table_to_fasta
+from phippery.phipdata import stitch_dataset
+from phippery.phipdata import collect_sample_table
+from phippery.phipdata import collect_peptide_table
 from phippery.phipdata import dataset_to_csv
-from phippery.phipdata import csv_to_dataset
-from phippery.phipdata import df_to_dataset
 from phippery.phipdata import load
 from phippery.phipdata import dump
 from phippery.phipdata import add_stats
+from phippery.phipdata import trim_index
+
+
+# TODO simulate a dataset directly to test these, it'll be a lot faster if we keep things
+# in memory rather than reading in the simulation tests each time.
+def test_convert_peptide_table_to_fasta(shared_datadir, tmp_path):
+    """
+    Test convert to fasta from peptide metadata.
+
+    Given a unique identifier and a set of Oligos, we expect the fasta file
+    to have a header ">" and Oligo sequence for each.
+    """
+
+    for sim_test in iter_sim_tests(shared_datadir):
+
+        d = tmp_path / "conv"
+        d.mkdir()
+        p = d / "test_fasta"
+        convert_peptide_table_to_fasta(sim_test.pep_meta, p)
+        assert os.path.exists(p)
+        pep_meta = sim_test.pds.peptide_table
+        for i, line in enumerate(open(p, "r")):
+            line = line.strip()
+            if i % 2 == 0:
+                assert line.startswith(">")
+                assert int(line[1:]) == pep_meta.loc[i // 2, :].peptide_id
+            else:
+                pep_meta_oligo = trim_index(str(pep_meta.loc[i // 2, "Oligo"].values))
+                assert line == pep_meta_oligo
+
+
+def test_trim_index():
+    """
+    Test trimming function
+    """
+
+    assert trim_index("aaaTTTaaa") == "TTT"
+    assert trim_index("") == ""
+    assert trim_index("TTT") == "TTT"
+    assert trim_index("aaaTTT") == "TTT"
+    assert trim_index("TTTaaa") == "TTT"
 
 
 def test_add_stats(shared_datadir, tmp_path):
@@ -77,11 +117,12 @@ def test_sims_generator(shared_datadir):
         assert type(sim_test.counts) == list
 
 
-def test_counts_metadata_to_dataset(shared_datadir):
+def test_stitch_dataset(shared_datadir):
 
     for sim_test in iter_sim_tests(shared_datadir):
-        pds = counts_metadata_to_dataset(
-            sim_test.counts, sim_test.pep_meta, sim_test.sam_meta
+
+        pds = stitch_dataset(
+            sim_test.xr_pd.counts.to_pandas(), sim_test.pep_meta, sim_test.sam_meta
         )
         assert type(pds) == xr.Dataset
         assert np.all(pds.counts.values == sim_test.solution)
@@ -100,7 +141,7 @@ def test_read_write_csv(shared_datadir, tmp_path):
         assert os.path.exists(f"{d}/test_sample_table.csv")
         assert os.path.exists(d / "test_peptide_table.csv")
 
-        ds_from_csv = csv_to_dataset(
+        ds_from_csv = stitch_dataset(
             f"{d}/test_counts.csv",
             f"{d}/test_peptide_table.csv",
             f"{d}/test_sample_table.csv",
@@ -113,7 +154,7 @@ def test_read_write_csv(shared_datadir, tmp_path):
 def test_df_to_dataset(shared_datadir):
 
     for sim_test in iter_sim_tests(shared_datadir):
-        ds = df_to_dataset(
+        ds = stitch_dataset(
             counts_df=sim_test.pds.counts.to_pandas(),
             peptide_table_df=sim_test.pds.peptide_table.to_pandas(),
             sample_table_df=sim_test.pds.sample_table.to_pandas(),
